% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata_to_Robjects.R
\name{get.ms.output}
\alias{get.ms.output}
\title{Read in msms data,}
\usage{
get.ms.output(ms.file = NA, software = msms)
}
\arguments{
\item{ms.file}{A string}

\item{software}{A string}
}
\value{
A list containing the simulation command 
physical positions and genotype sparse matrix from \code{ms.file}
}
\description{
Read in msms data,
}
\examples{
get.ms.output(ms.file, software="ms")
}
